pyproject.toml
[tool.poetry]
name = "tks-volsu-schedule"
version = "1.0.0"
description = "Бот, показывающий расписание для студентов, преподавателей и кабинетов кафедры ТКС ВолГУ"
authors = ["N0rmalUser <vova.okunev25@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.12.0"
APScheduler = "^3.10.4"
asyncio = "^3.4.3"
lets_plot = "^4.4.1"
matplotlib = "^3.9.2"
openpyxl = "^3.1.5"
pandas = "^2.2.2"
progress = "^1.6"
python-docx = "^1.1.2"
pytz = "^2024.1"
toml = "^0.10.2"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
black = "~24.8.0"
isort = "~5.13.2"
pre-commit = "~3.8.0"

[tool.poetry.scripts]
format = "black . && isort ."
replace-sensitive = "python replace_sensitive_data.py"

[tool.black]
line-length = 99
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 99
multi_line_output = 3
skip_gitignore = true
skip_glob = ["**/data/*", "**/logs/*"]
src_paths = ["."]